requirements.txt
fastapi
uvicorn
python-multipart
email-validator
mongoengine
python-jose[cryptography]